Move tech data to dedicated subdirectory
