+++ apps/website/src/data/tech/network.ts
